+++ app/scripts/components/kwic/kwic.ts
@@ -10,12 +10,13 @@ import { KwicWordScope } from "./kwic-word"
 import { SelectWordEvent } from "@/statemachine/types"
 import { ApiKwic, Token } from "@/backend/types"
 import { StoreService } from "@/services/store"
-import { QueryParamSort, QueryResponse } from "@/backend/types/query"
+import { QueryParams, QueryParamSort, QueryResponse } from "@/backend/types/query"
 import { CorpusTransformed } from "@/settings/config-transformed.types"
 import { JQueryExtended, JQueryStaticExtended } from "@/jquery.types"
 import { loc } from "@/i18n"
 import { calculateHitsPicture, HitsPictureItem, isKwic, isLinkedKwic, massageData, Row } from "@/kwic/kwic"
 import { corpusSelection } from "@/corpora/corpus_listing"
+import { RelationsSentencesParams } from "@/backend/types/relations-sentences"
 
 type KwicController = IController & {
     // Bindings
@@ -29,7 +30,7 @@ type KwicController = IController & {
     page: number
     pageEvent: (page: number) => void
     hitsPerPage: number
-    params: any
+    params: QueryParams | RelationsSentencesParams
     response?: QueryResponse
     corpusOrder: string[]
     /** Current page of results. */
@@ -622,7 +623,7 @@ angular.module("korpApp").component("kwic", {
                     const next = getNextToken()
                     if (next) {
                         next.trigger("click")
-                        scrollToShowWord(next)
+                        next.get(0)?.scrollIntoView({ block: "nearest", inline: "nearest" })
                         // Return false to prevent default behavior
                         return false
                     }
@@ -694,28 +695,6 @@ angular.module("korpApp").component("kwic", {
 
                 return output
             }
-
-            function scrollToShowWord(word: JQLite) {
-                if (!word.length) return
-                const offset = 200
-
-                if (word.offset()!.top + word.height()! > window.scrollY + $(window).height()!) {
-                    $("html, body")
-                        .stop(true, true)
-                        .animate({ scrollTop: window.scrollY + offset })
-                } else if (word.offset()!.top < window.scrollY) {
-                    $("html, body")
-                        .stop(true, true)
-                        .animate({ scrollTop: window.scrollY - offset })
-                }
-
-                const area = $element.find(".table_scrollarea")
-                if (word.offset()!.left + word.width()! > area.offset()!.left + area.width()!) {
-                    area.stop(true, true).animate({ scrollLeft: area.scrollLeft()! + offset })
-                } else if (word.offset()!.left < area.offset()!.left) {
-                    area.stop(true, true).animate({ scrollLeft: area.scrollLeft()! - offset })
-                }
-            }
         },
     ],
 })
@@ -757,7 +736,7 @@ class SelectionManager {
 // Add download links for other formats, defined in
 // settings["download_formats"] (Jyrki Niemi <jyrki.niemi@helsinki.fi>
 // 2014-02-26/04-30)
-export function setDownloadLinks(params: string, result_data: { kwic: Row[]; corpus_order: string[] }): void {
+export function setDownloadLinks(params: any, result_data: { kwic: Row[]; corpus_order: string[] }): void {
     // If some of the required parameters are null, return without
     // adding the download links.
     if (!(params != null && result_data != null && result_data.corpus_order != null && result_data.kwic != null)) {