+++ app/scripts/components/wordpicture/word-picture-column.ts
@@ -1,26 +1,26 @@
 import angular, { IController } from "angular"
 import { html } from "@/util"
-import { AlignedApiRelation, WordPictureColumn } from "@/word-picture"
+import { MatchedRelation, WordPicture } from "@/word-picture"
 import { Lemgram } from "@/lemgram"
 import { RelationsSort } from "@/backend/types/relations"
 import { RootScope } from "@/root-scope.types"
 import { WordpicExampleTask } from "@/task/wordpic-example-task"
 
 type WordPictureColumnController = IController & {
-    column: WordPictureColumn
+    cssClass: string
+    items: MatchedRelation[]
     limit: string
     showWordClass: boolean
     sort: RelationsSort
     // Locals
-    rows: AlignedApiRelation[]
-    parseLemgram: (row: AlignedApiRelation) => { label: string; pos?: string; idx?: number }
-    onClickExample: (row: AlignedApiRelation) => void
+    parseLemgram: (row: MatchedRelation) => { label: string; pos?: string; idx?: number }
+    onClickExample: (row: MatchedRelation) => void
 }
 
 angular.module("korpApp").component("wordPictureColumn", {
     template: html`
-        <div class="lemgram_result float-left p-1" ng-class="$ctrl.column.config.css_class">
-            <table class="m-0 p-0">
+        <div class="lemgram_result float-left p-1" ng-class="$ctrl.cssClass">
+            <table class="m-0">
                 <tbody>
                     <tr ng-repeat="row in $ctrl.rows" ng-init="data = $ctrl.parseLemgram(row)">
                         <td class="px-1 text-right"><span class="enumerate"></span></td>
@@ -51,7 +51,8 @@ angular.module("korpApp").component("wordPictureColumn", {
         </div>
     `,
     bindings: {
-        column: "<",
+        cssClass: "<",
+        items: "<",
         limit: "<",
         showWordClass: "<",
         sort: "<",
@@ -63,25 +64,27 @@ angular.module("korpApp").component("wordPictureColumn", {
 
             $ctrl.$onChanges = (changes) => {
                 if (changes.limit?.currentValue) {
-                    $ctrl.rows = $ctrl.column.rows.slice(0, Number($ctrl.limit))
+                    $ctrl.rows = $ctrl.items.slice(0, Number($ctrl.limit))
                 }
             }
 
             $ctrl.parseLemgram = function (row) {
-                const [id] = row.other.split("|")
-                const prefix = row.depextra ? `${row.depextra} ` : ""
+                const [other] = row.other.split("|")
+                const prefix = row.prefix ? `${row.prefix} ` : ""
 
-                const lemgram = Lemgram.parse(id)
+                const lemgram = Lemgram.parse(other)
                 if (lemgram) {
-                    const concept = row.dep ? lemgram.form : "-"
                     return {
-                        label: prefix + concept,
+                        label: prefix + lemgram.form,
                         pos: lemgram.pos,
                         idx: lemgram.index,
                     }
                 }
 
-                return { label: prefix + id }
+                return {
+                    label: prefix + other,
+                    pos: row.otherpos.toLowerCase(),
+                }
             }
 
             $ctrl.onClickExample = function (row) {