+++ app/scripts/jquery.types.ts
@@ -1,9 +1,6 @@
 export type JQueryExtended<T = HTMLElement> = JQuery<T> & {
     // Defined in /app/lib/jquery.localize.js, a modified version of jquery-localize
     localize: () => JQueryExtended<T>
-    // Defined in jq_extensions
-    outerHTML: () => JQueryExtended<T>
-    localeKey: (key: string) => JQueryExtended<T>
 }
 
 export type JQueryStaticExtended = JQueryStatic & {