+++ app/scripts/search/extended-search.ts
@@ -5,7 +5,7 @@ import { AttributeOption } from "@/corpora/corpus-set"
 import { loc, locAttribute } from "@/i18n"
 
 /** Load attribute values from backend data as selector options. */
-export async function loadOptions(attr: AttributeOption, lang: string) {
+export async function loadOptions(attr: AttributeOption, lang: string): Promise<string[][]> {
     const name = attr.name
     const split = attr.type === "set"
 
@@ -13,6 +13,8 @@ export async function loadOptions(attr: AttributeOption, lang: string) {
     const corpora = corpusSelection.corpora
         .filter((corpus) => name in corpus.struct_attributes || name in corpus.attributes)
         .map((corpus) => corpus.id)
+
+    if (!corpora.length) return []
 
     const data = await getAttrValues(corpora, name, split)