+++ app/scripts/word-picture.ts
@@ -1,8 +1,7 @@
-import { compact, mapKeys } from "lodash"
-import { ApiRelation } from "./backend/types/relations"
-import settings from "./settings"
+import { isEqual, once } from "lodash"
+import { Relation } from "./backend/types/relations"
+import { getWordPictureConfig } from "./settings"
 import { WordPictureDef, WordPictureDefItem } from "./settings/app-settings.types"
-import { uniqDeep } from "./util"
 
 export type WordType = "word" | "lemgram"
 
@@ -20,6 +19,7 @@ export type WordPictureSectionHeading = { word: string; pos: string }
 export type WordPictureTable = {
     config: WordPictureDef
     index: number
+    columns: WordPictureColumn[]
     columnsBefore: WordPictureColumn[]
     columnsAfter: WordPictureColumn[]
     max: number
@@ -28,10 +28,10 @@ export type WordPictureTable = {
 export type WordPictureColumn = {
     config: WordPictureDefItem
     rel: string
-    rows: AlignedApiRelation[]
+    rows: MatchedRelation[]
 }
 
-export type AlignedApiRelation = ApiRelation & {
+export type MatchedRelation = Relation & {
     /** True if the search word matches `dep` instead of `head`. */
     reverse: boolean
     /** Copy of `head` or `dep`, whichever matches the search word. */
@@ -40,43 +40,54 @@ export type AlignedApiRelation = ApiRelation & {
     /** Copy of `head` or `dep`, whichever doesn't match the search word. */
     other: string
     otherpos: string
+    /** Copy of `depextra` if the search word matches `head` – a preposition or other string to show together with the related word */
+    prefix?: string
 }
 
 export class WordPicture {
-    readonly tagset = settings["word_picture_tagset"] || {}
-    readonly config = mapKeys(settings["word_picture_conf"] || {}, (_, long) => this.tagset[long].toUpperCase())
-    readonly items: AlignedApiRelation[]
-    readonly headings: WordPictureSectionHeading[]
-    protected data: WordPictureData
+    readonly config = getWordPictureConfig()
+    readonly items: Record<string, MatchedRelation[]> = {}
+    readonly headings: WordPictureSectionHeading[] = []
 
     constructor(
         readonly query: string,
         readonly type: WordType,
-        itemsRaw: ApiRelation[],
+        itemsRaw: Relation[],
     ) {
-        const convertItem = (item: ApiRelation): AlignedApiRelation | undefined => {
-            const { head, headpos, dep, deppos } = item
+        const convertItem = (item: Relation): MatchedRelation | undefined => {
+            const { head, headpos, dep, deppos, depextra } = item
             // For ordinary word search, include multi-word items beginning with the searched word
             const getMatch = (word: string) => (type == "word" ? word.replace(/_.*/, "") : word)
             if (query == getMatch(head))
-                return { ...item, reverse: false, match: head, matchpos: headpos, other: dep, otherpos: deppos }
+                return {
+                    ...item,
+                    reverse: false,
+                    match: head,
+                    matchpos: headpos,
+                    other: dep,
+                    otherpos: deppos,
+                    prefix: depextra,
+                }
             if (query == getMatch(dep))
                 return { ...item, reverse: true, match: dep, matchpos: deppos, other: head, otherpos: headpos }
             console.warn("Unmatched relations item", item)
         }
 
-        this.items = compact(itemsRaw.map(convertItem))
-
-        const headings = uniqDeep(this.items.map((item) => ({ word: item.match, pos: item.matchpos })))
-        this.headings = headings.filter((heading) => heading.pos in this.config)
+        for (const itemRaw of itemsRaw) {
+            const item = convertItem(itemRaw)
+            if (!item) continue
+            const { match, matchpos, rel, reverse } = item
+            // Store items by category
+            const id = this.getColumnId(matchpos, rel, reverse)
+            this.items[id] ??= []
+            this.items[id].push(item)
+            // Store distinct section headings
+            const heading = { word: match, pos: matchpos }
+            if (!this.headings.find((h) => isEqual(h, heading))) this.headings.push(heading)
+        }
     }
 
-    getData(): WordPictureData {
-        if (!this.data) this.data = this.buildData()
-        return this.data
-    }
-
-    protected buildData() {
+    getData: () => WordPictureData = once(() => {
         return this.headings.map((heading) => {
             const config = this.config[heading.pos]
             const tables: WordPictureTable[] = config.map((config, index) => {
@@ -84,28 +95,27 @@ export class WordPicture {
                 const columnsBefore: WordPictureColumn[] = []
                 const columnsAfter: WordPictureColumn[] = []
                 let bin = columnsBefore
-                for (const column of config) {
-                    if (column == "_") bin = columnsAfter
-                    else bin.push(this.getColumn(column, heading.pos))
+                for (const col of config) {
+                    if (col == "_") bin = columnsAfter
+                    else {
+                        const id = this.getColumnId(heading.pos, col.rel, !!col.field_reverse)
+                        const rows = this.items[id]
+                        if (rows) bin.push({ config: col, rel: col.rel, rows })
+                    }
                 }
-                const max = Math.max(...[...columnsBefore, ...columnsAfter].map((col) => col.rows.length))
-                return { config, index, columnsBefore, columnsAfter, max }
+                const columns = [...columnsBefore, ...columnsAfter]
+                const max = Math.max(0, ...columns.map((col) => col.rows.length))
+                return { config, index, columns, columnsBefore, columnsAfter, max }
             })
             const max = Math.max(...tables.map((table) => table.max))
             return { config, heading, tables, max }
         })
-    }
+    })
 
     getMaxColumnLength(): number {
         return Math.max(...this.getData().map((section) => section.max))
     }
 
-    /** Filter items matching a column config. */
-    getColumn(config: WordPictureDefItem, matchpos: string): WordPictureColumn {
-        const rel = this.tagset[config.rel].toUpperCase()
-        const rows = this.items.filter(
-            (item) => item.matchpos == matchpos && item.rel == rel && item.reverse == !!config.field_reverse,
-        )
-        return { config, rel, rows }
-    }
+    /** Get a string for the params that identify a word picture column */
+    getColumnId = (pos: string, rel: string, reverse: boolean) => `${pos}${reverse ? "+" : "-"}${rel}`
 }